requirements.txt
mkdocs
mkdocs-material
Jinja2
pyyaml
mkdocs-minify-plugin